Refactor python filestructure
