+++ GuiApp/snackAttackTrack.kv
@@ -547,7 +547,7 @@
                 color: 'white'
                 width: '48dp'
                 size_hint_x: None
-                on_press: root.displayaboutpopup()
+                on_press: root.displayAboutPopup()
             Label:
                 id: heading
                 text: 'Snack Attack Track'
