+++ GuiApp/widgets/popups/abortPopup.py
@@ -1,0 +1,5 @@
+from kivy.uix.popup import Popup
+
+class AboutPopup(Popup):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
