+++ GuiApp/widgets/popups/errorMessagePopup.py
@@ -1,0 +1,6 @@
+from kivy.uix.popup import Popup
+
+class ErrorMessagePopup(Popup):
+    def __init__(self, errorMessage: str, **kwargs):
+        super(ErrorMessagePopup, self).__init__(**kwargs)
+        self.ids.errorMessagePopupLabel.text = errorMessage
